requirements.txt
requests==2.21.0
python-dotenv==0.10.2
python-telegram-bot==11.1.0
google-api-python-client==1.7.9
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
PyDrive==1.3.1
